+++ examples/html/javascript/udp.js
@@ -1,6 +1,44 @@
 let REQUESTID = 0
 
-export function post (bytes, timeout) {
+export function broadcast (bytes) {
+  debug(bytes)
+
+  const rq = {
+    ID: nextID(),
+    wait: '500ms',
+    request: [...bytes]
+  }
+
+  const request = {
+    method: 'POST',
+    mode: 'cors',
+    cache: 'no-cache',
+    credentials: 'same-origin',
+    headers: { 'Content-Type': 'application/json' },
+    redirect: 'follow',
+    referrerPolicy: 'no-referrer',
+    body: JSON.stringify(rq)
+  }
+
+  return fetch('/udp/broadcast', request)
+    .then(response => {
+      switch (response.status) {
+        case 200:
+          return response.json()
+
+        default:
+          response.text().then(w => {
+            throw new Error(w)
+          })
+      }
+    })
+    .then(reply => {
+      debug2(reply.replies)
+      return reply.replies
+    })
+}
+
+export function send (bytes, timeout) {
   debug(bytes)
 
   const rq = {
@@ -20,7 +58,7 @@ export function post (bytes, timeout) {
     body: JSON.stringify(rq)
   }
 
-  return fetch('/udp', request)
+  return fetch('/udp/send', request)
     .then(response => {
       switch (response.status) {
         case 200: